+++ Gruntfile.js
@@ -15,6 +15,7 @@ module.exports = function ( grunt ) {
 				'*.js',
 				'modules/**/*.js',
 				'tests/**/*.js',
+				'docs/**/*.js',
 				'**/*.json',
 				'!node_modules/**',
 				'!vendor/**'